BaseError && auth-post-resultErrors

+++ backend/services/post-service.js
@@ -1,3 +1,4 @@
+const BaseError = require('../errors/base-error')
 const postModel = require("../models/post-model")
 const fileService = require('./file-service')
 
@@ -8,26 +9,29 @@ class PostService {
     }
     async create(post, picture) {
         const fileName = await fileService.save(picture)
-        const createdPost = await postModel.create({ ...post, picture: fileName })
+        const createdPost = await postModel.create({ ...post, picture: fileName || "No photo"})
         return createdPost
     }
     async delete(id) {
+        if (!id) {
+            throw BaseError.BadRequest("Id is not found deleting post by id")
+        }
         const deletedPost = await postModel.findByIdAndDelete(id)
         return deletedPost
     }
     async edit(id, post) {
         if (!id) {
-            throw new Error("Id not found")
+            throw BaseError.BadRequest("Id is not found editing post by id")
         }
         const editedPost = await postModel.findByIdAndUpdate(id, post, { new: true })
         return editedPost
     }
     async getById(id) {
         if (!id) {
-            throw new Error("Id not found getting by id")
+            throw BaseError.BadRequest("Id is not found getting post by id")
         }
-        const takedPostById = await postModel.findById(id)
-        return takedPostById
+        const takenPostById = await postModel.findById(id)
+        return takenPostById
     }
 }
